MAISI train_v1

+++ maisi/train_v1_utils.py
@@ -88,7 +88,7 @@ def create_data_loader(
                 RandSpatialCropd(keys=input_modality, roi_size=output_size),
                 RandFlipd(keys=input_modality, prob=0.5),
                 RandRotate90d(keys=input_modality, prob=0.5),
-                ToTensord(keys=input_modality, dtype="half"),
+                ToTensord(keys=input_modality, dtype="float"),
             ]
         )
 
@@ -116,7 +116,7 @@ def create_data_loader(
                 LoadImaged(keys=input_modality, image_only=True),
                 EnsureChannelFirstd(keys=input_modality, channel_dim='no_channel'),
                 RandSpatialCropd(keys=input_modality, roi_size=output_size),
-                ToTensord(keys=input_modality, dtype="half"),
+                ToTensord(keys=input_modality, dtype="float"),
             ]
         )
 
@@ -144,7 +144,7 @@ def create_data_loader(
                 LoadImaged(keys=input_modality, image_only=True),                
                 EnsureChannelFirstd(keys=input_modality, channel_dim='no_channel'),
                 RandSpatialCropd(keys=input_modality, roi_size=output_size),
-                ToTensord(keys=input_modality, dtype="half"),
+                ToTensord(keys=input_modality, dtype="float"),
             ]
         )
